working without docker
