+++ .circleci/triggers/requirements.txt
@@ -1,0 +1,2 @@
+click==8.1.3
+requests==2.28.1
