+++ .circleci/triggers/utils.py
@@ -1,21 +1,14 @@
-#!/usr/bin/env python3
-
-import json
 import logging
-from pprint import pprint
 from subprocess import PIPE, Popen
-
-import click
-import requests
 
 logging.basicConfig(
     format="%(levelname)-8s [%(asctime)s] %(message)s", level=logging.INFO
 )
-logger = logging.getLogger("cli")
+logger = logging.getLogger("utils")
 
 
-# execute shell command and capture the output
 def run_cmd(cmd, exit_on_error=True, msg=None):
+    """Execute shell command and capture the output"""
     logger.debug(msg if msg else f"< {cmd}")
     sp = Popen(cmd.split(" "), stdout=PIPE, stderr=PIPE)
     stdout, stderr = sp.communicate()
@@ -40,6 +33,7 @@ def run_cmd(cmd, exit_on_error=True, msg=None):
 
 
 def standardize_branch(branches):
+    """Return 'master' and 'release' for all detected master/release branches"""
     for branch in branches:
         if "master" in branch or "remotes/origin/HEAD" in branch:
             return "master"
@@ -49,6 +43,7 @@ def standardize_branch(branches):
 
 
 def ensure_tag_on_master(tag):
+    """Ensure tag has been merged before proceeding"""
     try:
         branches = run_cmd(
             f"git name-rev --name-only {tag}", exit_on_error=False
@@ -61,31 +56,3 @@ def ensure_tag_on_master(tag):
     if branch not in ("master", "release"):
         print(f"Commit not found on master, nor release branches: {branch}")
         exit(1)
-
-
-@click.command()
-@click.option("-t", "--git-tag", required=True)
-@click.option("-k", "--circle-api-key", envvar="CIRCLE_API_KEY", required=True)
-def cli(git_tag, circle_api_key):
-    ensure_tag_on_master(git_tag)
-    org = "AudiusProject"
-    project = "audius-protocol"
-    branch = "master"
-    data = {"branch": branch, "parameters": {"sdk_release_tag": git_tag}}
-    r = requests.post(
-        f"https://circleci.com/api/v2/project/github/{org}/{project}/pipeline",
-        headers={"Content-Type": "application/json"},
-        auth=(circle_api_key, ""),
-        data=json.dumps(data),
-        allow_redirects=True,
-    )
-    response = r.json()
-    pprint(response)
-
-    pipeline_number = response["number"]
-    url = f"https://app.circleci.com/pipelines/github/AudiusProject/audius-protocol/{pipeline_number}"
-    print(f"\n{url}")
-
-
-if __name__ == "__main__":
-    cli()
